terminado service, serviceean y serviceRemote, falta controller

+++ Gestion-txoko-lagunak/Gestion-txoko-lagunak-ejb/src/main/java/org/lagunak/txoko/categoria/CategoriaServiceRemote.java
@@ -17,5 +17,5 @@ public interface CategoriaServiceRemote {
 	
 	public Categoria create(Categoria categoria);
 	
-	public void delete (int codigo);
+	public void delete (long codigo);
 }
